+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## 0.1.2-dev
+## 0.2.0-dev
+
+**Breaking changes**:
+
+- Removed `Element.text` and `Element.children`, using `Element.content` instead.
+- content items that are not `List`, `Component`, `String`, `Node` or `BuildFn` will be converted to `String` (and to `Text`). 
+
+Updates:
 
 - Fix: `BuildContext.ancestors` did not include `Component`s.
 - Fix: classes were not updated when the new Element had no class.